% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint}
\alias{blueprint}
\title{Create a Blueprint}
\usage{
blueprint(...)
}
\arguments{
\item{...}{One or more step objects.}
}
\value{
A blueprint object.
}
\description{
A blueprint is a list of steps that are executed in order.
Type \code{step_} and TAB in RStudio to see all available steps.
}
\examples{
blueprint(
  step_preprocess(),
  step_pca(),
  step_dea_limma(),  # this comma is ok
)
}
