% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint_default}
\alias{blueprint_default}
\title{Default blueprint}
\usage{
blueprint_default()
}
\value{
A \code{glysmith_blueprint} object.
}
\description{
This blueprint contains the following steps:
\itemize{
\item step_preprocess(): Preprocess the data using \code{glyclean::auto_clean()}.
\item step_ident_overview(): Summarize the experiment using \code{glyexp::summarize_experiment()}.
\item step_pca(): Principal component analysis using \code{glystats::gly_pca()},
and plot the PCA using \code{glyvis::plot_pca()}.
\item step_dea(): Differential analysis using \code{glystats::gly_dea()}.
\item step_volcano(): Plot a volcano plot using \code{glyvis::plot_volcano()}.
\item step_enrich_go(): Perform GO enrichment analysis using \code{glystats::gly_enrich_go()}.
\item step_enrich_kegg(): Perform KEGG enrichment analysis using \code{glystats::gly_enrich_kegg()}.
\item step_enrich_reactome(): Perform Reactome enrichment analysis using \code{glystats::gly_enrich_reactome()}.
\item step_derive_traits(): Derive traits using \code{glydet::derive_traits()}.
\item step_dta(): Differential trait analysis using \code{glystats::gly_limma()}.
}
}
\examples{
blueprint_default()
}
