% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{br}
\alias{br}
\title{Create a Branch in a Blueprint}
\usage{
br(name, ...)
}
\arguments{
\item{name}{Branch name used as a prefix for outputs.}

\item{...}{One or more step objects.}
}
\value{
A branch object used inside \code{\link[=blueprint]{blueprint()}}.
}
\description{
Use \code{br()} to group steps that should run as an isolated branch with
namespaced outputs prefixed by \verb{<name>__}.
}
\examples{
blueprint(
  step_preprocess(),
  br("limma",
    step_dea_limma(),
    step_volcano()
  ),
  br("ttest",
    step_dea_ttest(),
    step_volcano()
  )
)
}
