% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{cast_exp}
\alias{cast_exp}
\alias{cast_plot}
\alias{cast_table}
\title{Get Plots or Tables from GlySmith Result}
\usage{
cast_exp(x)

cast_plot(x, name)

cast_table(x, name)
}
\arguments{
\item{x}{A glysmith result object.}

\item{name}{The name of the plot or table to get.}
}
\value{
A glyexp_experiment object, a ggplot object or a tibble.
}
\description{
Helper functions to get processed experiment, plots or tables from a glysmith result object.
Just a syntax sugar for \verb{$exp}, \verb{$plots$plot_name} and \verb{$tables$table_name}
elements of a glysmith result object, respectively.
}
\examples{
library(glyexp)
exp <- real_experiment
result <- forge_analysis(exp)
cast_exp(result)
cast_plot(result, "pca")
cast_table(result, "summary")

}
