% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge-analysis.R
\name{forge_analysis}
\alias{forge_analysis}
\title{Forge an Analysis for Group Comparison}
\usage{
forge_analysis(exp, blueprint = blueprint_default(), group_col = "group", ...)
}
\arguments{
\item{exp}{A \code{glyexp::experiment()} object.}

\item{blueprint}{A \code{glysmith_blueprint} object. Default is \code{\link[=blueprint_default]{blueprint_default()}}.}

\item{group_col}{Column name of group information in the sample information.
Used for various analyses. Default is "group".}

\item{...}{Additional arguments passed to the underlying functions.
Use the format \code{step_id.pkg.func.arg} (step-scoped).
For example, if you want to pass argument \code{p_adj_method = "BH"} to \code{glystats::gly_limma()}
in \code{\link[=step_dea]{step_dea()}}, set \code{dea.glystats.gly_limma.p_adj_method = "BH"}.
To pass \code{batch_col} to \code{glyclean::auto_clean()} in \code{\link[=step_preprocess]{step_preprocess()}},
set \code{preprocess.glyclean.auto_clean.batch_col = "batch"}.
Note that arguments about group column specification is controlled by \code{group_col} argument,
and should not be passed to \code{...}.}
}
\value{
A \code{glysmith_result} object, with the following components:
\itemize{
\item \code{exp}: the experiment after preprocessing.
\item \code{plots}: a named list of ggplot objects.
\item \code{tables}: a named list of tibbles.
\item \code{meta}: a named list of metadata. Currently two elements:
\itemize{
\item \code{explanation}: a named character vector or list of explanations for each plot and table,
with keys like \code{tables$summary} and \code{plots$pca}.
\item \code{steps}: a character vector of the steps of the analysis.
}
}
}
\description{
This function performs a comprehensive analysis for group comparison.
}
\examples{
exp <- glyexp::real_experiment2
result <- forge_analysis(exp)
print(result)

}
