% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quench-result.R
\name{quench_result}
\alias{quench_result}
\title{Save GlySmith Result}
\usage{
quench_result(
  x,
  dir,
  plot_ext = "pdf",
  table_ext = "csv",
  plot_width = 5,
  plot_height = 5
)
}
\arguments{
\item{x}{A glysmith result object.}

\item{dir}{The directory to save the result.}

\item{plot_ext}{The extension of the plot files. Either "pdf", "png" or "svg". Default is "pdf".}

\item{table_ext}{The extension of the table files. Either "csv" or "tsv". Default is "csv".}

\item{plot_width}{The width of the plot in inches. Default is 5.}

\item{plot_height}{The height of the plot in inches. Default is 5.}
}
\value{
A glysmith result object.
}
\description{
Save processed experiment, plots and tables of a glysmith result object to a directory.
A \code{README.md} file will also be generated to describe the saved outputs,
based on metadata in \code{x$meta}.
}
\examples{
library(glyexp)
exp <- real_experiment
result <- forge_analysis(exp)
quench_result(result, tempdir())

}
