% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_dea}
\alias{step_dea}
\title{Step: Differential Expression Analysis (DEA)}
\usage{
step_dea()
}
\value{
A \code{glysmith_step} object.
}
\description{
Run differential analysis using \code{glystats::gly_limma()}.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to run DEA on
}

Data generated:
\itemize{
\item \code{dea_res}: The DEA results from \code{glystats::gly_limma()}
}

Tables generated:
\itemize{
\item \code{dea}: A table containing the DEA results
}
}
\examples{
step_dea()
}
\seealso{
\code{\link[glystats:gly_limma]{glystats::gly_limma()}}
}
