% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_dea_anova}
\alias{step_dea_anova}
\title{Step: Differential Expression Analysis (DEA) using ANOVA}
\usage{
step_dea_anova(...)
}
\arguments{
\item{...}{Step-specific arguments passed to \code{glystats::gly_anova()}.
Use the format \code{pkg.func.arg}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Run differential analysis using ANOVA via \code{glystats::gly_anova()}.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to run DEA on
}

Data generated:
\itemize{
\item \code{dea_res}: The DEA results from \code{glystats::gly_anova()}
}

Tables generated:
\itemize{
\item \code{dea_main_test}: A table containing the main test result
\item \code{dea_post_hoc_test}: A table containing the post-hoc test result
}
}
\examples{
step_dea_anova()
}
\seealso{
\code{\link[glystats:gly_anova]{glystats::gly_anova()}}
}
