% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_dea_limma}
\alias{step_dea_limma}
\title{Step: Differential Expression Analysis (DEA) using Limma}
\usage{
step_dea_limma(...)
}
\arguments{
\item{...}{Step-specific arguments passed to \code{glystats::gly_limma()}.
Use the format \code{pkg.func.arg}.
For example, \code{step_dea_limma(glystats.gly_limma.p_adj_method = "BH")}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Run differential analysis using linear model-based analysis via \code{glystats::gly_limma()}.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to run DEA on
}

Data generated:
\itemize{
\item \code{dea_res}: The DEA results from \code{glystats::gly_limma()}
}

Tables generated:
\itemize{
\item \code{dea}: A table containing the DEA result
}
}
\examples{
step_dea_limma()
}
\seealso{
\code{\link[glystats:gly_limma]{glystats::gly_limma()}}
}
