% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_derive_traits}
\alias{step_derive_traits}
\title{Step: Derived Trait Calculation}
\usage{
step_derive_traits(...)
}
\arguments{
\item{...}{Step-specific arguments passed to underlying functions.
Use the format \code{pkg.func.arg}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Calculate glycan derived traits using \code{glydet::derive_traits()}.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to calculate derived traits for
}

Data generated:
\itemize{
\item \code{trait_exp}: The experiment with derived traits
}

Tables generated:
\itemize{
\item \code{derived_traits}: A table containing the derived traits.
}
}
\examples{
step_derive_traits()
}
\seealso{
\code{\link[glydet:derive_traits]{glydet::derive_traits()}}
}
