% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_dta}
\alias{step_dta}
\title{Step: Differential Trait Analysis (DTA)}
\usage{
step_dta()
}
\value{
A \code{glysmith_step} object.
}
\description{
Run differential analysis on derived traits using \code{glystats::gly_limma()}.
This step requires \code{\link[=step_derive_traits]{step_derive_traits()}}.
}
\examples{
step_dta()
}
