% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_enrich}
\alias{step_enrich}
\title{Step: Enrichment Analysis}
\usage{
step_enrich(kind = c("go", "kegg", "reactome"), retry = 0L)
}
\arguments{
\item{kind}{Enrichment type: \code{"go"}, \code{"kegg"}, or \code{"reactome"}.}

\item{retry}{Number of retries if the step errors.}
}
\value{
A \code{glysmith_step} object.

A \code{glysmith_step} object.
}
\description{
Run functional enrichment analysis on differentially expressed variables using one of:
\itemize{
\item \code{glystats::gly_enrich_go()}
\item \code{glystats::gly_enrich_kegg()}
\item \code{glystats::gly_enrich_reactome()}
}

This step requires \code{\link[=step_dea]{step_dea()}}.
}
\examples{
step_enrich("go")
}
