% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_enrich_reactome}
\alias{step_enrich_reactome}
\title{Step: Reactome Enrichment Analysis}
\usage{
step_enrich_reactome()
}
\value{
A \code{glysmith_step} object.
}
\description{
Perform Reactome enrichment analysis on differentially expressed variables using \code{glystats::gly_enrich_reactome()}.
This step requires \code{\link[=step_dea]{step_dea()}}.
}
\examples{
step_enrich_reactome()
}
