% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_ident_overview}
\alias{step_ident_overview}
\title{Step: Identification Overview}
\usage{
step_ident_overview()
}
\value{
A \code{glysmith_step} object.
}
\description{
Summarize the experiment using \code{glyexp::summarize_experiment()}.
}
\examples{
step_ident_overview()
}
