% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_logo}
\alias{step_logo}
\title{Step: Logo Plot}
\usage{
step_logo(on = "exp", n_aa = 5L, fasta = NULL, ...)
}
\arguments{
\item{on}{Name of the experiment data in \code{ctx$data} to plot.
One of "exp", "sig_exp". Default is "exp".}

\item{n_aa}{The number of amino acids to the left and right of the glycosylation site.
For example, if \code{n_aa = 5}, the resulting sequence will contain 11 amino acids.}

\item{fasta}{Path to the FASTA file containing the amino acid sequences.
If \code{NULL}, this step will be skipped.}

\item{...}{Additional arguments passed to \code{\link[ggseqlogo:ggseqlogo]{ggseqlogo::ggseqlogo()}}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Create a logo plot for glycosylation sites using \code{glyvis::plot_logo()}.
The logo plot visualizes the amino acid sequence patterns around glycosylation sites.
This step is only applicable for glycoproteomics experiments.
}
\details{
Data required:
\itemize{
\item Depends on \code{on} parameter (default: \code{exp})
}

Plots generated:
\itemize{
\item \code{logo}: A logo plot (if \code{on = "exp"})
\item \code{sig_logo}: A logo plot (if \code{on = "sig_exp"})
}
}
\examples{
step_logo()
step_logo(fasta = "proteins.fasta")
step_logo(on = "sig_exp")
}
\seealso{
\code{\link[glyvis:plot_logo]{glyvis::plot_logo()}}
}
