% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-dimension.R
\name{step_pca}
\alias{step_pca}
\title{Step: Principal Component Analysis (PCA)}
\usage{
step_pca(
  on = "exp",
  center = TRUE,
  scale = TRUE,
  loadings = FALSE,
  screeplot = TRUE,
  plot_width = 5,
  plot_height = 5,
  ...
)
}
\arguments{
\item{on}{Name of the experiment to run PCA on.
Can be "exp", "sig_exp", "trait_exp", "sig_trait_exp",
"dynamic_motif_exp", "sig_dynamic_motif_exp",
"branch_motif_exp", "sig_branch_motif_exp".}

\item{center}{A logical indicating whether to center the data. Default is TRUE.}

\item{scale}{A logical indicating whether to scale the data. Default is TRUE.}

\item{loadings}{Logical indicating whether to generate the loading plot.
Default is \code{FALSE} since loading plots for glycoproteomics data can be crowded.}

\item{screeplot}{Logical indicating whether to generate the screeplot.
Default is \code{TRUE}.}

\item{plot_width}{Width of plots in inches. Default is 5.}

\item{plot_height}{Height of plots in inches. Default is 5.}

\item{...}{Additional arguments passed to \code{prcomp()}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Run PCA using \code{glystats::gly_pca()} and plot it with \code{glyvis::plot_pca()}.
Loading plot for glycoproteomics data can be crowded with too many variables.
Ignore the resulting plot if it is not informative.
}
\details{
Data required:
\itemize{
\item \code{exp} (if \code{on = "exp"}): The experiment to run PCA on
\item \code{trait_exp} (if \code{on = "trait_exp"}): The trait experiment to run PCA on
\item \code{dynamic_motif_exp} (if \code{on = "dynamic_motif_exp"}): The dynamic motif experiment to run PCA on
\item \code{branch_motif_exp} (if \code{on = "branch_motif_exp"}): The branch motif experiment to run PCA on
}

Tables generated (with suffixes):
\itemize{
\item \code{pca_samples}: A table containing the PCA scores for each sample
\item \code{pca_variables}: A table containing the PCA loadings for each variable
\item \code{pca_eigenvalues}: A table containing the PCA eigenvalues
}

Plots generated (with suffixes):
\itemize{
\item \code{pca_scores}: A PCA score plot colored by group (always generated)
\item \code{pca_loadings}: A PCA loading plot (if \code{loadings = TRUE})
\item \code{pca_screeplot}: A PCA screeplot (if \code{screeplot = TRUE})
}
}
\section{AI Prompt}{

\emph{This section is for AI in \code{\link[=inquire_blueprint]{inquire_blueprint()}} only.}
\itemize{
\item Include this step if needed.
}
}

\examples{
step_pca()
}
\seealso{
\code{\link[glystats:gly_pca]{glystats::gly_pca()}}, \code{\link[glyvis:plot_pca]{glyvis::plot_pca()}}
}
