% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_quantify_motifs}
\alias{step_quantify_motifs}
\title{Step: Quantify Motifs}
\usage{
step_quantify_motifs(...)
}
\arguments{
\item{...}{Step-specific arguments passed to underlying functions.
Use the format \code{pkg.func.arg}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Quantify glycan motifs using \code{glydet::quantify_motifs()}.
The motifs are extracted using \code{glymotif::extract_branch_motif()} for N-glycans
and \code{glymotif::extract_motif()} for others.
Advanced glycan motif analysis that quantify glycan motifs (substructures) of a glycome or each glycosite.
Need glycan structure information.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to quantify motifs for
}

Data generated:
\itemize{
\item \code{motif_exp}: The experiment with quantified motifs
}

Tables generated:
\itemize{
\item \code{quantified_motifs}: A table containing the quantified motifs.
}
}
\section{Dynamic Arguments}{
This step supports the following dynamic arguments:
\itemize{
\item \code{glymotif.extract_motif.max_size}: Maximum size of motifs to extract.
\item \code{glydet.quantify_motifs.method}: "relative" or "absolute".
}
}

\examples{
step_quantify_motifs()
}
\seealso{
\code{\link[glydet:quantify_motifs]{glydet::quantify_motifs()}}, \code{\link[glymotif:extract_motif]{glymotif::extract_motif()}}, \code{\link[glymotif:extract_branch_motif]{glymotif::extract_branch_motif()}}
}
