% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_roc}
\alias{step_roc}
\title{Step: ROC Analysis}
\usage{
step_roc(...)
}
\arguments{
\item{...}{Step-specific arguments passed to \code{glystats::gly_roc()} and \code{glyvis::plot_roc()}.
Use the format \code{pkg.func.arg}.
For example, \code{step_roc(glystats.gly_roc.pos_class = "positive_class")}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Perform ROC analysis using \code{glystats::gly_roc()},
extract top 10 variables with highest AUC,
and plot ROC curves for these variables using \code{glyvis::plot_roc()}.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to perform ROC analysis on
}

Tables generated:
\itemize{
\item \code{roc_auc}: A table containing the ROC AUC values for all variables
}

Plots generated:
\itemize{
\item \code{roc_curves}: ROC curves for the top 10 variables
}
}
\section{Dynamic Arguments}{
This step supports the following dynamic arguments:
\itemize{
\item \code{glystats.gly_roc.pos_class}: The positive class.
}
}

\examples{
step_roc()
}
\seealso{
\code{\link[glystats:gly_roc]{glystats::gly_roc()}}, \code{\link[glyvis:plot_roc]{glyvis::plot_roc()}}
}
