% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_sig_enrich_reactome}
\alias{step_sig_enrich_reactome}
\title{Step: Reactome Enrichment Analysis on Differentially Expressed Variables}
\usage{
step_sig_enrich_reactome(universe = "all", ...)
}
\arguments{
\item{universe}{The universe (background) to use for enrichment analysis.
One of "all" (all genes in OrgDb), "detected" (detected variables in \code{exp}).}

\item{...}{Step-specific arguments passed to underlying functions.
Use the format \code{pkg.func.arg}.
For example, \code{step_sig_enrich_reactome(glystats.gly_enrich_reactome.OrgDb = "org.Mm.eg.db")}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Perform Reactome enrichment analysis on differentially expressed variables using \code{glystats::gly_enrich_reactome()}.
This step requires one of the DEA steps to be run.
Only execute for glycoproteomics experiments with exactly 2 groups.
If used for glycomics experiments, the step will be skipped.
Use all genes in OrgDb as the background.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to perform Reactome enrichment analysis for
\item \code{sig_exp}: The filtered experiment from DEA, generated by \code{step_dea_xxx()}.
}

Tables generated:
\itemize{
\item \code{reactome_enrich}: A table containing the Reactome enrichment results.
}
}
\section{Dynamic Arguments}{
This step supports the following dynamic arguments:
\itemize{
\item \code{glystats.gly_enrich_reactome.OrgDb}: Organism database (default: "org.Hs.eg.db").
\item \code{glystats.gly_enrich_reactome.organism}: Organism name (default: "human").
\item \code{glyvis.plot_enrich.type}: Plot type ("dotplot", "barplot", etc.).
}
}

\examples{
step_sig_enrich_reactome()
}
\seealso{
\code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_reactome()}}
}
