% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-dimension.R
\name{step_tsne}
\alias{step_tsne}
\title{Step: t-SNE}
\usage{
step_tsne(on = "exp", dims = 2, perplexity = 30, ...)
}
\arguments{
\item{on}{Name of the experiment to run t-SNE on.
Can be "exp", "sig_exp", "trait_exp", "sig_trait_exp", "motif_exp", "sig_motif_exp".}

\item{dims}{Number of output dimensions. Default is 2.}

\item{perplexity}{Perplexity parameter for t-SNE. Default is 30.}

\item{...}{Additional arguments passed to \code{Rtsne::Rtsne()}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Perform t-SNE analysis using \code{glystats::gly_tsne()} and
plot a t-SNE plot using \code{glyvis::plot_tsne()}.
Note that the result of t-SNE largely depends on the \code{perplexity} parameter.
Usually it's a trial-and-error process to find the best value iteratively.
If you are not satisfied with the result,
manually call \code{glyvis::plot_tsne()} with different \code{perplexity} values
to find the best one.
}
\details{
Data required:
\itemize{
\item \code{exp} (if \code{on = "exp"}): The experiment to perform t-SNE on
\item \code{trait_exp} (if \code{on = "trait_exp"}): The trait experiment to perform t-SNE on
\item \code{motif_exp} (if \code{on = "motif_exp"}): The motif experiment to perform t-SNE on
}

Data generated (with suffixes):
\itemize{
\item \code{tsne}: The t-SNE result
}

Plots generated (with suffixes):
\itemize{
\item \code{tsne}: The t-SNE plot
}
}
\section{AI Prompt}{

\emph{This section is for AI in \code{\link[=inquire_blueprint]{inquire_blueprint()}} only.}
\itemize{
\item Include this step only when the user explicitly asks for t-SNE.
}
}

\examples{
step_tsne()
step_tsne(perplexity = 30)
}
\seealso{
\code{\link[glystats:gly_tsne]{glystats::gly_tsne()}}, \code{\link[glyvis:plot_tsne]{glyvis::plot_tsne()}}
}
