% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_volcano}
\alias{step_volcano}
\title{Step: Volcano Plot}
\usage{
step_volcano()
}
\value{
A \code{glysmith_step} object.
}
\description{
Create a volcano plot from DEA results using \code{glyvis::plot_volcano()}.
This step requires \code{\link[=step_dea]{step_dea()}}.
}
\examples{
step_volcano()
}
