% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{write_blueprint}
\alias{write_blueprint}
\alias{read_blueprint}
\title{Save or Load a Blueprint}
\usage{
write_blueprint(bp, file)

read_blueprint(file)
}
\arguments{
\item{bp}{A \code{\link[=blueprint]{blueprint()}}.}

\item{file}{A character string giving the name of the file to save to or load from.}
}
\value{
Invisibly returns the blueprint object.
}
\description{
\itemize{
\item \code{write_blueprint()} saves a blueprint to a RDS file.
\item \code{read_blueprint()} loads a blueprint from a RDS file.
}
}
\examples{
bp <- blueprint(
  step_preprocess(),
  step_pca(),
  step_dea(),
)
write_blueprint(bp, tempfile(fileext = ".rds"))
}
