% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint_default}
\alias{blueprint_default}
\title{Default blueprint}
\usage{
blueprint_default(preprocess = TRUE, enrich = TRUE, traits = TRUE)
}
\arguments{
\item{preprocess}{Whether to include \code{\link[=step_preprocess]{step_preprocess()}}.}

\item{enrich}{Whether to include the enrichment steps,
i.e. \code{\link[=step_sig_enrich_go]{step_sig_enrich_go()}}, \code{\link[=step_sig_enrich_kegg]{step_sig_enrich_kegg()}}, and \code{\link[=step_sig_enrich_reactome]{step_sig_enrich_reactome()}}.}

\item{traits}{Whether to include the derived trait analysis steps,
i.e. \code{\link[=step_derive_traits]{step_derive_traits()}} and \code{step_dea_limma(on = "trait_exp")}.}
}
\value{
A \code{glysmith_blueprint} object.
}
\description{
This blueprint contains the following steps:
\itemize{
\item step_ident_overview(): Summarize the experiment using \code{glyexp::summarize_experiment()}.
\item step_preprocess(): Preprocess the data using \code{glyclean::auto_clean()}.
\item step_plot_qc(when = "post"): Plot QC plots using \code{glyclean::plot_qc()}.
\item step_pca(): Principal component analysis using \code{glystats::gly_pca()},
and plot the PCA using \code{glyvis::plot_pca()}.
\item step_dea_limma(): Differential analysis using \code{glystats::gly_limma()}.
\item step_volcano(): Plot a volcano plot using \code{glyvis::plot_volcano()}.
\item step_heatmap(on = "sig_exp"): Plot a heatmap using \code{glyvis::plot_heatmap()}.
\item step_sig_enrich_go(): Perform GO enrichment analysis using \code{glystats::gly_enrich_go()}.
\item step_sig_enrich_kegg(): Perform KEGG enrichment analysis using \code{glystats::gly_enrich_kegg()}.
\item step_sig_enrich_reactome(): Perform Reactome enrichment analysis using \code{glystats::gly_enrich_reactome()}.
\item step_derive_traits(): Derive traits using \code{glydet::derive_traits()}.
\item step_dea_limma(on = "trait_exp"): Differential trait analysis using \code{glystats::gly_limma()}.
\item step_heatmap(on = "sig_trait_exp"): Plot a heatmap using \code{glyvis::plot_heatmap()}.
}
}
\examples{
blueprint_default()
}
