% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glysmith-result.R
\name{cast_exp}
\alias{cast_exp}
\alias{cast_plot}
\alias{cast_table}
\alias{cast_data}
\title{Get Data from GlySmith Result}
\usage{
cast_exp(x)

cast_plot(x, name)

cast_table(x, name)

cast_data(x, name)
}
\arguments{
\item{x}{A glysmith result object.}

\item{name}{The name of the plot or table to get.}
}
\value{
\itemize{
\item \code{cast_exp()}: a \code{\link[glyexp:experiment]{glyexp::experiment()}}.
\item \code{cast_plot()}: a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
\item \code{cast_table()}: a \code{\link[tibble:tibble]{tibble::tibble()}}.
\item \code{cast_data()}: can be any R object.
}
}
\description{
Helper functions to get processed experiment, plots, tables or data from a glysmith result object.
}
\examples{
library(glyexp)
exp <- real_experiment2
result <- forge_analysis(exp)
cast_exp(result)
cast_table(result, "summary")

}
