% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge-analysis.R
\name{forge_analysis}
\alias{forge_analysis}
\title{Forge an Analysis for Group Comparison}
\usage{
forge_analysis(exp, blueprint = blueprint_default(), group_col = "group", ...)
}
\arguments{
\item{exp}{A \code{glyexp::experiment()} object.}

\item{blueprint}{A \code{glysmith_blueprint} object. Default is \code{\link[=blueprint_default]{blueprint_default()}}.}

\item{group_col}{Column name of group information in the sample information.
Used for various analyses. Default is "group".}

\item{...}{Additional arguments passed to the functions.
Use the format \code{pkg.func.arg} to pass arguments to the functions.
For example, if you want to pass argument \code{p_adj_method = "BH"} to \code{glystats::gly_limma()},
set \code{glystats.gly_limma.p_adj_method = "BH"}.
Note that arguments about group column specification is controlled by \code{group_col} argument,
and should not be passed to \code{...}.}
}
\value{
A \code{glysmith_result} object, with the following components:
\itemize{
\item \code{exp}: the experiment after preprocessing.
\item \code{plots}: a named list of ggplot objects.
\item \code{tables}: a named list of tibbles.
\item \code{meta}: a named list of metadata. Currently two elements:
\itemize{
\item \code{explanation}: a named character vector or list of explanations for each plot and table,
with keys like \code{tables$summary} and \code{plots$pca}.
\item \code{steps}: a character vector of the steps of the analysis.
}
}
}
\description{
This function performs comprehensive analysis for group comparison.

The pipeline includes:
\itemize{
\item Preprocessing using \code{glyclean::auto_clean()}
\item Identification overview using \code{glyexp::summarize_experiment()}
\item Principal component analysis (PCA) with \code{glystats::gly_pca()} and \code{glyvis::plot_pca()}
\item Differential expression analysis (DEA) with \code{glystats::gly_limma()} and \code{glyvis::plot_volcano()}
}

If experiment type is "glycoproteomics", the pipeline will also include:
\itemize{
\item Functional enrichment analysis using \code{glystats::gly_enrich_go()}, \code{glystats::gly_enrich_kegg()},
and \code{glystats::gly_enrich_reactome()}, as well as plotting with \code{glyvis::plot_enrich()}
}

If glycan structure is available and glycan type is "N",
the pipeline will also include:
\itemize{
\item Derived trait calculation using \code{glydet::derive_traits()}
\item Differential trait analysis (DTA) with \code{glystats::gly_limma()}
}
}
\examples{
exp <- glyexp::real_experiment2
result <- forge_analysis(exp)
print(result)

}
