% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inquire-blueprint.R
\name{inquire_blueprint}
\alias{inquire_blueprint}
\title{Create a Blueprint using Natural Language}
\usage{
inquire_blueprint(
  description,
  exp = NULL,
  group_col = "group",
  model = "deepseek-reasoner",
  max_retries = 3
)
}
\arguments{
\item{description}{A description of what you want to analysis.}

\item{exp}{Optional. A \code{glyexp::experiment()} object to provide more context to the LLM.}

\item{group_col}{The column name of the group variable in the experiment. Default to "group".}

\item{model}{Model to use. Default to "deepseek-reasoner".}

\item{max_retries}{Maximum number of retries when the AI output is invalid. Default to 3.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Ask a Large Language Model (LLM) to create a blueprint for glycomics or glycoproteomics data analysis.
To use this function, you need to have a DeepSeek API key.
You can get a DeepSeek API key from https://platform.deepseek.com.
Then set the environment variable \code{DEEPSEEK_API_KEY} to your API key with
\code{Sys.setenv(DEEPSEEK_API_KEY = "your-api-key")}.
}
\details{
LLMs can be unstable. If you get an error, try again with another description.
Make sure to examine the returned blueprint carefully to ensure it's what you want.
You can also create parallel analysis branches with \code{br("name", step_..., step_...)},
which will namespace outputs with the branch prefix.

Here are some examples that works:
\itemize{
\item "I want to know what pathways are enriched for my differentially expressed glycoforms."
\item "I want a heatmap and a pca plot. I have already performed preprocessing myself."
\item "I have a glycomics dataset. I want to calculate derived traits and perform DEA on them."
}
}
