% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_adjust_protein}
\alias{step_adjust_protein}
\title{Step: Adjust Protein Abundance}
\usage{
step_adjust_protein(pro_expr_path = NULL, method = "ratio")
}
\arguments{
\item{pro_expr_path}{Path to the protein expression matrix file.
If \code{NULL}, this step will be skipped.
Can be:
\itemize{
\item A CSV/TSV file with the first column as protein accessions and remaining columns as sample names.
\item An RDS file with a matrix or data.frame with row names as protein accessions and columns as sample names.
}}

\item{method}{The method to use for protein adjustment.
Either "ratio" or "reg". Default is "ratio".}
}
\value{
A \code{glysmith_step} object.
}
\description{
Adjust glycoform quantification values by correcting for protein abundance
utilizing \code{glyclean::adjust_protein()}.
Usually this step should be run after \code{step_preprocess()}.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to adjust
}

Data generated:
\itemize{
\item \code{unadj_exp}: The original experiment (previous \code{exp}, saved for reference)
}

This step is special in that it silently overwrites the \code{exp} data with the adjusted experiment.
This ensures that no matter if adjustment is performed or not,
the "active" experiment is always under the key \code{exp}.
The previous \code{exp} is saved as \code{unadj_exp} for reference.
}
\examples{
step_adjust_protein("protein_expr.csv")
step_adjust_protein("protein_expr.rds", method = "reg")
}
\seealso{
\code{\link[glyclean:adjust_protein]{glyclean::adjust_protein()}}
}
