% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_dea}
\alias{step_dea}
\title{Step: Differential Expression Analysis (DEA)}
\usage{
step_dea()
}
\value{
A \code{glysmith_step} object.
}
\description{
Run differential analysis using \code{glystats::gly_limma()}.
}
\examples{
step_dea()
}
