% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_dea_kruskal}
\alias{step_dea_kruskal}
\title{Step: Differential Expression Analysis (DEA) using Kruskal-Wallis test}
\usage{
step_dea_kruskal(...)
}
\arguments{
\item{...}{Step-specific arguments passed to \code{glystats::gly_kruskal()}.
Use the format \code{pkg.func.arg}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Run differential analysis using Kruskal-Wallis analysis via \code{glystats::gly_kruskal()}.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to run DEA on
}

Data generated:
\itemize{
\item \code{dea_res}: The DEA results from \code{glystats::gly_kruskal()}
}

Tables generated:
\itemize{
\item \code{dea_main_test}: A table containing the main test result
\item \code{dea_post_hoc_test}: A table containing the post-hoc test result
}
}
\examples{
step_dea_kruskal()
}
\seealso{
\code{\link[glystats:gly_kruskal]{glystats::gly_kruskal()}}
}
