% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_dea_ttest}
\alias{step_dea_ttest}
\title{Step: Differential Expression Analysis (DEA) using t-test}
\usage{
step_dea_ttest(...)
}
\arguments{
\item{...}{Step-specific arguments passed to \code{glystats::gly_ttest()}.
Use the format \code{pkg.func.arg}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Run differential analysis using t-test via \code{glystats::gly_ttest()}.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to run DEA on
}

Data generated:
\itemize{
\item \code{dea_res}: The DEA results from \code{glystats::gly_ttest()}
}

Tables generated:
\itemize{
\item \code{dea}: A table containing the DEA result
}
}
\examples{
step_dea_ttest()
}
\seealso{
\code{\link[glystats:gly_ttest]{glystats::gly_ttest()}}
}
