% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_dea_wilcox}
\alias{step_dea_wilcox}
\title{Step: Differential Expression Analysis (DEA) using Wilcoxon test}
\usage{
step_dea_wilcox(...)
}
\arguments{
\item{...}{Step-specific arguments passed to \code{glystats::gly_wilcox()}.
Use the format \code{pkg.func.arg}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Run differential analysis using Wilcoxon analysis via \code{glystats::gly_wilcox()}.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to run DEA on
}

Data generated:
\itemize{
\item \code{dea_res}: The DEA results from \code{glystats::gly_wilcox()}
}

Tables generated:
\itemize{
\item \code{dea}: A table containing the DEA result
}
}
\examples{
step_dea_wilcox()
}
\seealso{
\code{\link[glystats:gly_wilcox]{glystats::gly_wilcox()}}
}
