% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_derive_traits}
\alias{step_derive_traits}
\title{Step: Derived Trait Calculation}
\usage{
step_derive_traits(...)
}
\arguments{
\item{...}{Step-specific arguments passed to underlying functions.
Use the format \code{pkg.func.arg}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Calculate glycan derived traits using \code{glydet::derive_traits()}.
Advanced glycan structure analysis that summarizes structural properties of a glycome or each glycosite.
Need glycan structure information.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to calculate derived traits for
}

Data generated:
\itemize{
\item \code{trait_exp}: The experiment with derived traits
}

Tables generated:
\itemize{
\item \code{derived_traits}: A table containing the derived traits.
}
}
\section{Dynamic Arguments}{
This step supports the following dynamic arguments:
\itemize{
\item \code{glydet.derive_traits.trait_fns}: Custom trait functions to calculate.
\item \code{glydet.derive_traits.mp_fns}: Custom meta-property functions.
\item \code{glydet.derive_traits.mp_cols}: Column names to use as meta-properties.
}
}

\examples{
step_derive_traits()
}
\seealso{
\code{\link[glydet:derive_traits]{glydet::derive_traits()}}
}
