% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_derive_traits}
\alias{step_derive_traits}
\title{Step: Derived Trait Calculation}
\usage{
step_derive_traits()
}
\value{
A \code{glysmith_step} object.
}
\description{
Calculate glycan derived traits using \code{glydet::derive_traits()}.
}
\examples{
step_derive_traits()
}
