% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_derive_traits}
\alias{step_derive_traits}
\title{Step: Derived Trait Calculation}
\usage{
step_derive_traits(trait_fns = NULL, mp_fns = NULL, mp_cols = NULL)
}
\arguments{
\item{trait_fns}{A named list of derived trait functions created by trait factories.
Names of the list are the names of the derived traits.
Default is \code{NULL}, which means all derived traits in \code{\link[glydet:basic_traits]{basic_traits()}} are calculated.}

\item{mp_fns}{A named list of meta-property functions.
This parameter is useful if your trait functions use custom meta-properties
other than those in \code{\link[glydet:all_mp_fns]{all_mp_fns()}}.
Default is \code{NULL}, which means all meta-properties in \code{\link[glydet:all_mp_fns]{all_mp_fns()}} are used.}

\item{mp_cols}{A character vector of column names in the \code{var_info} tibble to use as meta-properties.
If names are provided, they will be used as names of the meta-properties,
otherwise the column names will be used.
Meta-properties specified in \code{mp_cols} will overwrite those introduced by \code{mp_fns} with the same names,
including the built-in meta-properties.
Default is \code{NULL}, which means no columns are used as meta-properties.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Calculate glycan derived traits using \code{glydet::derive_traits()}.
Advanced glycan structure analysis that summarizes structural properties of a glycome or each glycosite.
Need glycan structure information.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to calculate derived traits for
}

Data generated:
\itemize{
\item \code{trait_exp}: The experiment with derived traits
}

Tables generated:
\itemize{
\item \code{derived_traits}: A table containing the derived traits.
}
}
\examples{
step_derive_traits()
}
\seealso{
\code{\link[glydet:derive_traits]{glydet::derive_traits()}}
}
