% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_dta}
\alias{step_dta}
\title{Step: Differential Trait Analysis (DTA)}
\usage{
step_dta(...)
}
\arguments{
\item{...}{Step-specific arguments passed to underlying functions.
Use the format \code{pkg.func.arg}.
For example, \code{step_dta(glystats.gly_limma.p_adj_method = "BH")}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Run differential analysis on derived traits using \code{glystats::gly_limma()}.
This step requires \code{\link[=step_derive_traits]{step_derive_traits()}}.
}
\details{
Data required:
\itemize{
\item \code{trait_exp}: The experiment with derived traits
}

Tables generated:
\itemize{
\item \code{dta_res}: A table containing the differential trait analysis results.
}
}
\examples{
step_dta()
}
\seealso{
\code{\link[glystats:gly_limma]{glystats::gly_limma()}}
}
