% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_enrich_kegg}
\alias{step_enrich_kegg}
\title{Step: KEGG Enrichment Analysis}
\usage{
step_enrich_kegg()
}
\value{
A \code{glysmith_step} object.
}
\description{
Perform KEGG enrichment analysis on differentially expressed variables using \code{glystats::gly_enrich_kegg()}.
This step requires \code{\link[=step_dea]{step_dea()}}.
}
\examples{
step_enrich_kegg()
}
