% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_ident_overview}
\alias{step_ident_overview}
\title{Step: Identification Overview}
\usage{
step_ident_overview(...)
}
\arguments{
\item{...}{Step-specific arguments passed to underlying functions.
Use the format \code{pkg.func.arg}.
For example, \code{step_ident_overview(glyexp.summarize_experiment.count_struct = FALSE)}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Summarize the experiment using \code{glyexp::summarize_experiment()}.
This step can be run at any time, but is usually run before or right after preprocessing.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to summarize
}

Tables generated:
\itemize{
\item \code{summary}: A table containing the identification overview of the experiment
}
}
\examples{
step_ident_overview()
}
\seealso{
\code{\link[glyexp:count_compositions]{glyexp::summarize_experiment()}}
}
