% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_ident_overview}
\alias{step_ident_overview}
\title{Step: Identification Overview}
\usage{
step_ident_overview(...)
}
\arguments{
\item{...}{Step-specific arguments passed to underlying functions.
Use the format \code{pkg.func.arg}.
For example, \code{step_ident_overview(glyexp.summarize_experiment.count_struct = FALSE)}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Summarize the experiment using \code{glyexp::summarize_experiment()}.
This is usually the first step, BEFORE \code{step_preprocess()}.
Very light-weight to run, so always include it.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to summarize
}

Tables generated:
\itemize{
\item \code{summary}: A table containing the identification overview of the experiment
}
}
\section{Dynamic Arguments}{
This step supports the following dynamic arguments:
\itemize{
\item \code{glyexp.summarize_experiment.count_struct}: Whether to count by structure or composition.
}
}

\examples{
step_ident_overview()
}
\seealso{
\code{\link[glyexp:summarize_experiment]{glyexp::summarize_experiment()}}
}
