% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_ident_overview}
\alias{step_ident_overview}
\title{Step: Identification Overview}
\usage{
step_ident_overview(count_struct = NULL)
}
\arguments{
\item{count_struct}{For counting glycopeptides and glycoforms.
whether to count the number of glycan structures or glycopeptides.
If \code{TRUE}, glycopeptides or glycoforms bearing different glycan structures
with the same glycan composition are counted as different ones.
If not provided (NULL), defaults to \code{TRUE} if \code{glycan_structure} column exists
in the variable information tibble, otherwise \code{FALSE}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Summarize the experiment using \code{glyexp::summarize_experiment()}.
This is usually the first step, BEFORE \code{step_preprocess()}.
Very light-weight to run, so always include it.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to summarize
}

Tables generated:
\itemize{
\item \code{summary}: A table containing the identification overview of the experiment
}
}
\section{AI Prompt}{

\emph{This section is for AI in \code{\link[=inquire_blueprint]{inquire_blueprint()}} only.}
\itemize{
\item Always include this step by default unless the user explicitly excludes it.
\item Use it as the first step in the blueprint.
}
}

\examples{
step_ident_overview()
}
\seealso{
\code{\link[glyexp:summarize_experiment]{glyexp::summarize_experiment()}}
}
