% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_pca}
\alias{step_pca}
\title{Step: Principal Component Analysis (PCA)}
\usage{
step_pca(...)
}
\arguments{
\item{...}{Step-specific arguments passed to underlying functions.
Use the format \code{pkg.func.arg}.
For example, \code{step_pca(glystats.gly_pca.center = FALSE)}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Run PCA using \code{glystats::gly_pca()} and plot it with \code{glyvis::plot_pca()}.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to run PCA on
}

Tables generated:
\itemize{
\item \code{pca_samples}: A table containing the PCA scores for each sample
\item \code{pca_variables}: A table containing the PCA loadings for each variable
\item \code{pca_eigenvalues}: A table containing the PCA eigenvalues
}

Plots generated:
\itemize{
\item \code{pca}: A PCA plot colored by group
}
}
\section{Dynamic Arguments}{
This step supports the following dynamic arguments:
\itemize{
\item \code{glystats.gly_pca.center}: Whether to center the data (default: TRUE).
\item \code{glystats.gly_pca.scale}: Whether to scale the data (default: TRUE).
\item \code{glyvis.plot_pca.type}: Plot type ("screeplot", "individual", "variables", "biplot").
\item \code{glyvis.plot_pca.groups}: Group membership for coloring.
}
}

\examples{
step_pca()
}
\seealso{
\code{\link[glystats:gly_pca]{glystats::gly_pca()}}, \code{\link[glyvis:plot_pca]{glyvis::plot_pca()}}
}
