% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_pca}
\alias{step_pca}
\title{Step: Principal Component Analysis (PCA)}
\usage{
step_pca(on = "exp", center = TRUE, scale = TRUE, ...)
}
\arguments{
\item{on}{Name of the experiment to run PCA on.
Can be "exp", "sig_exp", "trait_exp", "sig_trait_exp", "motif_exp", "sig_motif_exp".}

\item{center}{A logical indicating whether to center the data. Default is TRUE.}

\item{scale}{A logical indicating whether to scale the data. Default is TRUE.}

\item{...}{Additional arguments passed to \code{prcomp()}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Run PCA using \code{glystats::gly_pca()} and plot it with \code{glyvis::plot_pca()}.
Loading plot for glycoproteomics data can be crowded with too many variables.
Ignore the resulting plot if it is not informative.
}
\details{
Data required:
\itemize{
\item \code{exp} (if \code{on = "exp"}): The experiment to run PCA on
\item \code{trait_exp} (if \code{on = "trait_exp"}): The trait experiment to run PCA on
\item \code{motif_exp} (if \code{on = "motif_exp"}): The motif experiment to run PCA on
}

Tables generated (with suffixes):
\itemize{
\item \code{pca_samples}: A table containing the PCA scores for each sample
\item \code{pca_variables}: A table containing the PCA loadings for each variable
\item \code{pca_eigenvalues}: A table containing the PCA eigenvalues
}

Plots generated (with suffixes):
\itemize{
\item \code{pca_scores}: A PCA score plot colored by group
\item \code{pca_loadings}: A PCA loading plot
\item \code{pca_screeplot}: A PCA screeplot
}
}
\examples{
step_pca()
}
\seealso{
\code{\link[glystats:gly_pca]{glystats::gly_pca()}}, \code{\link[glyvis:plot_pca]{glyvis::plot_pca()}}
}
