% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_pca}
\alias{step_pca}
\title{Step: Principal Component Analysis (PCA)}
\usage{
step_pca()
}
\value{
A \code{glysmith_step} object.
}
\description{
Run PCA using \code{glystats::gly_pca()} and plot it with \code{glyvis::plot_pca()}.
}
\examples{
step_pca()
}
