% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_preprocess}
\alias{step_preprocess}
\title{Step: Preprocessing}
\usage{
step_preprocess()
}
\value{
A \code{glysmith_step} object.
}
\description{
Preprocess the experiment using \code{glyclean::auto_clean()}.
}
\examples{
step_preprocess()
}
