% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_preprocess}
\alias{step_preprocess}
\title{Step: Preprocessing}
\usage{
step_preprocess(...)
}
\arguments{
\item{...}{Step-specific arguments passed to underlying functions.
Use the format \code{pkg.func.arg}.
For example, \code{step_preprocess(glyclean.auto_clean.remove_preset = "discovery")}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Preprocess the experiment using \code{glyclean::auto_clean()}.
This is usually the first step, but can be omitted if the experiment is already preprocessed.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to preprocess
}

Data generated:
\itemize{
\item \code{raw_exp}: The raw experiment (previous \code{exp}, saved for reference)
}

This step is special in that it silently overwrites the \code{exp} data with the preprocessed experiment.
This ensures that no matter if preprocessing is performed or not,
the "active" experiment is always under the key \code{exp}.
The previous \code{exp} is saved as \code{raw_exp} for reference.
}
\section{Dynamic Arguments}{
This step supports the following dynamic arguments:
\itemize{
\item \code{glyclean.auto_clean.batch_col}: Column name for batch information.
\item \code{glyclean.auto_clean.qc_name}: Name of QC samples (default: "QC").
\item \code{glyclean.auto_clean.normalize_to_try}: List of normalization functions to try.
\item \code{glyclean.auto_clean.impute_to_try}: List of imputation functions to try.
\item \code{glyclean.auto_clean.remove_preset}: Preset for removing variables ("simple", "discovery", "biomarker").
\item \code{glyclean.auto_clean.batch_prop_threshold}: Proportion threshold for batch correction (default: 0.3).
\item \code{glyclean.auto_clean.check_batch_confounding}: Whether to check batch confounding (default: TRUE).
\item \code{glyclean.auto_clean.batch_confounding_threshold}: Threshold for batch confounding (default: 0.4).
}
}

\examples{
step_preprocess()
}
\seealso{
\code{\link[glyclean:auto_clean]{glyclean::auto_clean()}}
}
