% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_sig_enrich_go}
\alias{step_sig_enrich_go}
\title{Step: GO Enrichment Analysis on Differentially Expressed Variables}
\usage{
step_sig_enrich_go(universe = "all")
}
\arguments{
\item{universe}{The universe (background) to use for enrichment analysis.
One of "all" (all genes in OrgDb), "detected" (detected variables in \code{exp}).}
}
\value{
A \code{glysmith_step} object.
}
\description{
Perform GO enrichment analysis on differentially expressed variables using \code{glystats::gly_enrich_go()}.
This step requires \code{\link[=step_dea]{step_dea()}}.
Only execute for glycoproteomics experiments.
Use all genes in OrgDb as the background.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to perform GO enrichment analysis for
\item \code{dea_res}: The DEA results from \code{glystats::gly_limma()}
}

Tables generated:
\itemize{
\item \code{go_enrich}: A table containing the GO enrichment results.
}
}
\examples{
step_sig_enrich_go()
}
\seealso{
\code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_go()}}
}
