% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_sig_enrich_kegg}
\alias{step_sig_enrich_kegg}
\title{Step: KEGG Enrichment Analysis on Differentially Expressed Variables}
\usage{
step_sig_enrich_kegg(universe = "all", ...)
}
\arguments{
\item{universe}{The universe (background) to use for enrichment analysis.
One of "all" (all genes in OrgDb), "detected" (detected variables in \code{exp}).}

\item{...}{Step-specific arguments passed to underlying functions.
Use the format \code{pkg.func.arg}.
For example, \code{step_sig_enrich_kegg(glystats.gly_enrich_kegg.p_adj_method = "BH")}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Perform KEGG enrichment analysis on differentially expressed variables using \code{glystats::gly_enrich_kegg()}.
This step requires \code{\link[=step_dea]{step_dea()}}.
Only execute for glycoproteomics experiments.
Use all genes in OrgDb as the background.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to perform KEGG enrichment analysis for
\item \code{dea_res}: The DEA results from \code{glystats::gly_limma()}
}

Tables generated:
\itemize{
\item \code{kegg_enrich}: A table containing the KEGG enrichment results.
}
}
\examples{
step_sig_enrich_kegg()
}
\seealso{
\code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_kegg()}}
}
