% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_sig_enrich_kegg}
\alias{step_sig_enrich_kegg}
\title{Step: KEGG Enrichment Analysis on Differentially Expressed Variables}
\usage{
step_sig_enrich_kegg(universe = "all", ...)
}
\arguments{
\item{universe}{The universe (background) to use for enrichment analysis.
One of "all" (all genes in OrgDb), "detected" (detected variables in \code{exp}).}

\item{...}{Step-specific arguments passed to underlying functions.
Use the format \code{pkg.func.arg}.
For example, \code{step_sig_enrich_kegg(glystats.gly_enrich_kegg.p_adj_method = "BH")}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Perform KEGG enrichment analysis on differentially expressed variables using \code{glystats::gly_enrich_kegg()}.
This step requires one of the DEA steps to be run.
Only execute for glycoproteomics experiments with exactly 2 groups.
Use all genes in OrgDb as the background.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to perform KEGG enrichment analysis for
\item \code{sig_exp}: The filtered experiment from DEA, generated by \code{step_dea_xxx()}.
}

Tables generated:
\itemize{
\item \code{kegg_enrich}: A table containing the KEGG enrichment results.
}
}
\section{Dynamic Arguments}{
This step supports the following dynamic arguments:
\itemize{
\item \code{glystats.gly_enrich_kegg.OrgDb}: Organism database (default: "org.Hs.eg.db").
\item \code{glyvis.plot_enrich.type}: Plot type ("dotplot", "barplot", etc.).
}
}

\examples{
step_sig_enrich_kegg()
}
\seealso{
\code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_kegg()}}
}
