% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_sig_enrich_kegg}
\alias{step_sig_enrich_kegg}
\title{Step: KEGG Enrichment Analysis on Differentially Expressed Variables}
\usage{
step_sig_enrich_kegg(universe = "all", plot_type = "dotplot", ...)
}
\arguments{
\item{universe}{The universe (background) to use for enrichment analysis.
One of "all" (all genes in OrgDb), "detected" (detected variables in \code{exp}).}

\item{plot_type}{Plot type for enrichment results ("dotplot", "barplot", etc.).}

\item{...}{Additional arguments passed to \code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_kegg()}}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Perform KEGG enrichment analysis on differentially expressed variables using \code{glystats::gly_enrich_kegg()}.
This step requires one of the DEA steps to be run.
Only execute for glycoproteomics experiments with exactly 2 groups.
If used for glycomics experiments, the step will be skipped.
Use all genes in OrgDb as the background.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to perform KEGG enrichment analysis for
\item \code{sig_exp}: The filtered experiment from DEA, generated by \code{step_dea_xxx()}.
}

Tables generated:
\itemize{
\item \code{kegg_enrich}: A table containing the KEGG enrichment results.
}
}
\examples{
step_sig_enrich_kegg()
step_sig_enrich_kegg(plot_type = "barplot")
}
\seealso{
\code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_kegg()}}
}
