% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_tsne}
\alias{step_tsne}
\title{Step: t-SNE}
\usage{
step_tsne(on = "exp", ...)
}
\arguments{
\item{on}{Name of the experiment to run t-SNE on. Can be "exp", "sig_exp", "trait_exp", or "sig_trait_exp".}

\item{...}{Step-specific arguments passed to \code{glystats::gly_tsne()} and \code{glyvis::plot_tsne()}.
Use the format \code{pkg.func.arg}.
For example, \code{step_tsne(glystats.gly_tsne.perplexity = 30)}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Perform t-SNE analysis using \code{glystats::gly_tsne()} and
plot a t-SNE plot using \code{glyvis::plot_tsne()}.
Note that the result of t-SNE largely depends on the \code{perplexity} parameter.
Usually it's a trial-and-error process to find the best value iteratively.
If you are not satisfied with the result,
manually call \code{glyvis::plot_tsne()} with different \code{perplexity} values
to find the best one.
}
\details{
Data required:
\itemize{
\item \code{exp}: The experiment to perform t-SNE on
}

Data generated:
\itemize{
\item \code{tsne}: The t-SNE result
}

Plots generated:
\itemize{
\item \code{tsne}: The t-SNE plot
}
}
\section{Dynamic Arguments}{
This step supports the following dynamic arguments:
\itemize{
\item \code{glystats.gly_tsne.perplexity}: The perplexity parameter for t-SNE.
\item \code{glystats.gly_tsne.dims}: The number of dimensions for t-SNE.
\item \code{glystats.gly_tsne.xxx}: xxx are other parameters of \code{Rtsne::Rtsne()}.
}
}

\examples{
step_tsne()
step_tsne(glystats.gly_tsne.perplexity = 30)
}
\seealso{
\code{\link[glystats:gly_tsne]{glystats::gly_tsne()}}, \code{\link[glyvis:plot_tsne]{glyvis::plot_tsne()}}
}
