% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_umap}
\alias{step_umap}
\title{Step: UMAP}
\usage{
step_umap(on = "exp", ...)
}
\arguments{
\item{on}{Name of the experiment to run UMAP on.
Can be "exp", "sig_exp", "trait_exp", "sig_trait_exp", "motif_exp", "sig_motif_exp".}

\item{...}{Step-specific arguments passed to \code{glystats::gly_umap()} and \code{glyvis::plot_umap()}.
Use the format \code{pkg.func.arg}.
For example, \code{step_umap(glystats.gly_umap.n_neighbors = 15)}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Perform UMAP analysis using \code{glystats::gly_umap()} and
plot a UMAP plot using \code{glyvis::plot_umap()}.
Note that the result of UMAP largely depends on the \code{n_neighbors} parameter.
Usually it's a trial-and-error process to find the best value iteratively.
If you are not satisfied with the result,
manually call \code{glyvis::plot_umap()} with different \code{n_neighbors} values
to find the best one.
}
\details{
Data required:
\itemize{
\item \code{exp} (if \code{on = "exp"}): The experiment to perform UMAP on
\item \code{trait_exp} (if \code{on = "trait_exp"}): The trait experiment to perform UMAP on
\item \code{motif_exp} (if \code{on = "motif_exp"}): The motif experiment to perform UMAP on
}

Data generated (with suffixes):
\itemize{
\item \code{umap}: The UMAP result
}

Plots generated (with suffixes):
\itemize{
\item \code{umap}: The UMAP plot
}
}
\section{Dynamic Arguments}{
This step supports the following dynamic arguments:
\itemize{
\item \code{glystats.gly_umap.n_neighbors}: The number of neighbors.
\item \code{glystats.gly_umap.n_components}: The number of dimensions.
\item \code{glystats.gly_umap.xxx}: xxx are other parameters of \code{uwot::umap()}.
}
}

\examples{
step_umap()
step_umap(glystats.gly_umap.n_neighbors = 15)
}
\seealso{
\code{\link[glystats:gly_umap]{glystats::gly_umap()}}, \code{\link[glyvis:plot_umap]{glyvis::plot_umap()}}
}
