% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_volcano}
\alias{step_volcano}
\title{Step: Volcano Plot}
\usage{
step_volcano(...)
}
\arguments{
\item{...}{Step-specific arguments passed to underlying functions.
Use the format \code{pkg.func.arg}.
For example, \code{step_volcano(glyvis.plot_volcano.log2fc_cutoff = 2)}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Create a volcano plot from DEA results using \code{glyvis::plot_volcano()}.
This step requires one of the DEA steps to be run:
\itemize{
\item \code{\link[=step_dea_limma]{step_dea_limma()}}
\item \code{\link[=step_dea_ttest]{step_dea_ttest()}}
\item \code{\link[=step_dea_wilcox]{step_dea_wilcox()}}
}
}
\details{
Data required:
\itemize{
\item \code{dea_res}: The DEA results from \code{glystats::gly_limma()}
}

Plots generated:
\itemize{
\item \code{volcano}: A volcano plot
}
}
\section{Dynamic Arguments}{
This step supports the following dynamic arguments:
\itemize{
\item \code{glyvis.plot_volcano.log2fc_cutoff}: Log2 fold change cutoff (default: 1).
\item \code{glyvis.plot_volcano.p_cutoff}: P-value cutoff (default: 0.05).
\item \code{glyvis.plot_volcano.p_col}: Column for p-value ("p_adj" or "p_val").
\item \code{glyvis.plot_volcano.contrast}: Contrast to plot for limma results.
}
}

\examples{
step_volcano()
}
\seealso{
\code{\link[glyvis:plot_volcano]{glyvis::plot_volcano()}}
}
