% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step_volcano}
\alias{step_volcano}
\title{Step: Volcano Plot}
\usage{
step_volcano(log2fc_cutoff = 1, p_cutoff = 0.05, p_col = "p_adj", ...)
}
\arguments{
\item{log2fc_cutoff}{The log2 fold change cutoff. Defaults to 1.}

\item{p_cutoff}{The p-value cutoff. Defaults to 0.05.}

\item{p_col}{The column name for p-value. Defaults to "p_adj".
Can also be "p_val" (raw p-values without multiple testing correction).}

\item{...}{Other arguments passed to \code{\link[EnhancedVolcano:EnhancedVolcano]{EnhancedVolcano::EnhancedVolcano()}}.}
}
\value{
A \code{glysmith_step} object.
}
\description{
Create a volcano plot from DEA results using \code{glyvis::plot_volcano()}.
This step requires one of the DEA steps to be run:
\itemize{
\item \code{\link[=step_dea_limma]{step_dea_limma()}} (multi-group comparison is also supported)
\item \code{\link[=step_dea_ttest]{step_dea_ttest()}}
\item \code{\link[=step_dea_wilcox]{step_dea_wilcox()}}
}
}
\details{
Data required:
\itemize{
\item \code{dea_res}: The DEA results from \code{glystats::gly_limma()}
}

Plots generated:
\itemize{
\item \code{volcano}: A volcano plot
}
}
\section{AI Prompt}{

\emph{This section is for AI in \code{\link[=inquire_blueprint]{inquire_blueprint()}} only.}
\itemize{
\item Always include this step by default if DEA is performed, and the DEA method is not ANOVA or Kruskal-Wallis.
}
}

\examples{
step_volcano()
step_volcano(log2fc_cutoff = 2)
}
\seealso{
\code{\link[glyvis:plot_volcano]{glyvis::plot_volcano()}}
}
